Urban prefabs + some hacks
- Hitting backspace in prefab mode will delete all voxels on the map
that are in the prefab's cube. This needs to be an option on painting
prefabs: "erase voxels first" or something like that. Only it will also
need to check for the actual bounds of the sprite (min and max x,y,z of
voxels within the sprite size) rather than the entire cube. I know what
I mean anyway!
- Fixed rotation of prefabs